Move the plugin_list files to the plugins_common target.
The PluginList class depends on platform specific implementations of the ReadWebPluginInfo
function. This function was defined and implemented by the PluginLib class which iivs in the
plugin target. I moved this function to the PluginList class and deleted the platform specific
incarnations of the PluginLib class which only existed for this functionality.

Next step would be to link with plugin_common only in the chrome_browser target and to move the
common plugin files to the webkit\common folder.

BUG=237249
R=jamesr@chromium.org

Review URL: https://codereview.chromium.org/15345002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@200946 0039d316-1c4b-4281-b951-d872f2087c98