Allow pages on the OSX app launcher to be turned while dragging.
Adds an interface to the AppsGridController to optionally trigger a
page scroll after a delay, when given a drag location. Scrolling is
triggered if the point is to either side of the grid view, or over a
pager segment.

BUG=138633
TEST=Added: app_list_unittests AppsGridControllerTest.ScrollingWhileDragging.
Manually, test by dragging items on the OSX app launcher past the left
or right side, or over a pager segment, to initiate page scrolling
after a 1 second delay.

Review URL: https://chromiumcodereview.appspot.com/14999013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@200987 0039d316-1c4b-4281-b951-d872f2087c98