Refactoring: removed an unnessesary dependency to CLD (Compact Langauge Detector).
For UMA, we need to get an integer which identifies an language code, and CLD was used for this purpose.  This is not a primary usage of CLD.  Instead, I added a new function to get an integer from a language code with a simple algorithm.  Added a new metrics LanguageUsage.* instead of LanguageUsageMetrics.*.  LanguageUsageMetrics.* and enum Language will be deprecated.

BUG=177647
TSET=unit_tests

Review URL: https://chromiumcodereview.appspot.com/15262002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201056 0039d316-1c4b-4281-b951-d872f2087c98