Create an app list shim for the App Launcher on OSX in response to a startup flag.
This allows an app launcher .app shim to be created on OSX with steps:
 - Quit Chrome, and
 - $ open Chromium.app --args --enable-app-list-shim

Chrome will start, and a finder window will be opened showing the newly
created .app. This can then be launched and pinned to the dock.

The .app can be deleted with --enable-app-list-shim=0 (or it can be
dragged to Trash).

BUG=138633
TEST=Steps as above.

Review URL: https://chromiumcodereview.appspot.com/14760002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201260 0039d316-1c4b-4281-b951-d872f2087c98