[Autofill] Grant user consent for geolocation requests on Wallet dialog acceptance.
Geolocation includes network requests to Google servers, so the user needs to
consent to the collection of this information prior to us being able to collect
it.  The Autofill dialog (in Wallet mode) includes a notice informing the user
that location data will be collected for Risk purposes, so acceptance
(submission) of the dialog grants this consent.

BUG=166596,173505
TEST=none

Review URL: https://chromiumcodereview.appspot.com/14819016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201285 0039d316-1c4b-4281-b951-d872f2087c98