Simple Cache: Eliminate Open+Create race with no index.
With this change a failed Open returns to UNINITIALIZED state so that we later can perform the real operation and not "fail early just because we are in the failed state". Failed Create also lands as UNINITIALIZED, that's the closest state to having no synchronous entry.

BUG=239223

Review URL: https://chromiumcodereview.appspot.com/14858045

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201373 0039d316-1c4b-4281-b951-d872f2087c98