Starting to remove tests that might not be really essential. Explanations for why a particular test is being removed are below.
Please correct if the reasoning is not, uh, correct and also provide the reason(s) for doing so. 

For seek tests: 
Retain only WiFi and NoConstraints network conditions. Cable may be removed because it has a higher bandwidth (5120kbps) than WiFi (1024kbps) 
Remove BUFFERED seek tests, because the CACHED seeks perform the same actions.
This reduces the # of seek test cases from 180 to 64. 

For media-stats test: 
Retain only the 2160 and 1080 resolutions. 2160 is the highest and should be most likely to catch regressions. 1080 gives us something against which to compare performance, and is the most-expected usage.

This reduces the # of media-stats tests from 11 to 8, and # of metrics calculated from 44 to 32. 

BUG=240899

Review URL: https://chromiumcodereview.appspot.com/14932017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201706 0039d316-1c4b-4281-b951-d872f2087c98