Optimize TilingData::TileBoundsWithBorder().
TileBoundsWithBorder called TileBounds(), then had a bunch of if
statements that undid all of the if statements in TileBounds. It's
more efficient just to do the computation directly, without the
function call and with far fewer conditionals; in infinite scrolling
test pages (http://groupcloned.com/test/) we see typically 100
calls WithBorder and 200 without border in the compositor on every
frame, as well as a few dozen calls of each on the renderer and
rasterizer threads.

This seems to provide a 1-2% speedup across most of cc_perftests
on N7-class hardware.

BUG=

Review URL: https://chromiumcodereview.appspot.com/15896002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@201942 0039d316-1c4b-4281-b951-d872f2087c98