Parametrize names of llc and ld nexes by reading them from the resource info JSON file.
At Chrome build-time, the pnacl_component_crx_gen.py script generates a "resource info" JSON file that's placed alongside the translator nexes. We were not actually reading this file so far, but with this CL, we are. The first use is not hard-coding the ld.nexe and llc.nexe tool names in the plugin code. Instead, the tool names are read from the info file. This will allow us, in the future, to move the tool names into the native client directory so we can change them without touching Chrome at all.


BUG=None

Review URL: https://chromiumcodereview.appspot.com/15697019

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@202002 0039d316-1c4b-4281-b951-d872f2087c98