Add HasWord(string) method to SpellcheckCustomDictionary
This CL adds a method HasWord(string) to SpellcheckCustomDictionary and changes
the internal data structure in SpellcheckCustomDictionary from a vector to a set
for more efficient querying. Spelling service client will use the new method to
send IN_DICTIONARY feedback to spelling service for words that are in user's
custom spellcheck dictionary. Spelling service client will use this method in
upcoming CLs.

This CL also modifies IPC message "SpellCheckMsg_Init" to pass a set instead of
a vector of custom dictionary words. The browser sends this message to the
renderer to initialize spellcheck.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/15940004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@202722 0039d316-1c4b-4281-b951-d872f2087c98