[Telemetry] Don't restart the browser when the WPR archive path changes.
Simply restart WPR using the same port it was on before. The browser will still
be set up to point to that port.

I think we originally restarted the browser because we used to not explicitly
pass the ports to use to WPR. Now that those are stored in browser_backend,
they can stay stable across WPR restarts.

BUG=None
TEST=tools/perf/run_measurement --browser=android-content-shell page_cycler tools/perf/page_sets/top_25.json --page-filter="youtube|plus" --pageset-repeat=1 -v
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/15917016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@202877 0039d316-1c4b-4281-b951-d872f2087c98