Instant Extended: Suppress mouse events for bottom web content
When the omnibox overlay is open both the overlay's web contents and
the active tab's web contents get mouse events.

This causes problems like the cursor changing due to links below the
omnibox overlay.

A similar problem was fixed for the overlapping find bar by asvitkine
in crrev.com/6676094. This CL piggybacks on that fix by also checking
for overlapping web contents.

BUG=238846
TEST=Navigate to google.com. Type "http://youtube.com" in the omnibox.
Mouse over the top of the overlay. Without my fix the cursor changes
due to the links on google.com. With my fix the cursor does not change.

Review URL: https://chromiumcodereview.appspot.com/15315005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@203222 0039d316-1c4b-4281-b951-d872f2087c98