SimpleCache optimistic writes should not keep references to their IO buffer.
After the write operation returns OK, the caller is free to reuse the IOBuffer.
It is invalid to do, as the SimpleCache did previously, return OK and keep a
reference and use the IOBuffer.

As it happens, the BufferResource handler does reuse IO buffers, and this was
causing cache read errors due to CRC mismatches from the IO buffer being reused.

R=pasko,felipeg
BUG=239223

Review URL: https://chromiumcodereview.appspot.com/15825012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@203477 0039d316-1c4b-4281-b951-d872f2087c98