gpu: Avoid flinging beyond content edge
If the user flings down a page while the page is still loading, we may
hit a situation where the fling animation stops at the bottom edge of
the content, pauses while the loading proceeds and then continues once the
content grows taller. This results in a jarring pause in an otherwise
continuous animation.

This patch fixes the issue by preventing the fling animation from trying
to scroll beyond the content edge. In other words, as soon as the
content overscrolls on a specific axis, we prevent the fling animation
from scrolling on that axis again.

BUG=244376
TEST=InputHandlerProxyTest.GestureFlingStopsAtContentEdge

Review URL: https://chromiumcodereview.appspot.com/16156002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204006 0039d316-1c4b-4281-b951-d872f2087c98