Use hash_map for syncable::Directory indices
The hash map is less convenient to use, but it promises much better
expected algorithmic complexity for the most common directory
operations. The indices converted with this patch almost never require
iteration, so it makes much more sense to use a hash_map (ie.
unordered_map) than a regular map.

This change also introduces a new index based on the server-assigned
unique tag. The server-assigned tag is set only on permanent items, so
there won't be many items in this index.

I ran some tests locally to confirm the new code does run faster.  On
certain model association tests I saw a speedup of ~10%.  I doubt we'll
see that much of a speedup in real world scenarios, but it could be a
noticeable improvement on some of our slower platforms.

BUG=238621, 115132, 241813

Review URL: https://chromiumcodereview.appspot.com/16231009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204069 0039d316-1c4b-4281-b951-d872f2087c98