Add tools/android/memdump/.
This adds the 'memdump' binary that dumps device memory information for a list
of processes and 'memreport.py' (executed on the host) used to process this
information.

Memdump is similar to /proc/<PID>/smaps except that it is multi-process aware.
For a list of processes (e.g. browser + renderer processes), it is able to
dinstinguish the memory pages that are mapped only in these processes from the
ones that are private or also mapped in external processes.

This is useful to gather high level memory usage information (e.g. heap size vs
code size) for any (non-instrumented) process.

R=bulach@chromium.org, digit@chromium.org, qsr@chromium.org

Review URL: https://codereview.chromium.org/15861023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@204254 0039d316-1c4b-4281-b951-d872f2087c98